+++ src/lib/components/decode-message/cosmos-message/decode-message-object-transfer.tsx
@@ -47,14 +47,8 @@ export const DecodeMessageObjectTransfer = ({
   const nftObject = {
     collectionAddress: zAddr.optional().parse(nftMetadata?.collectionAddress),
     nftAddress: zHexAddr32.parse(formatAddresses(object).hex),
-    tokenId:
-      nftMetadata?.tokenId && typeof nftMetadata.tokenId === "string"
-        ? nftMetadata.tokenId
-        : undefined,
-    uri:
-      nftMetadata?.tokenUri && typeof nftMetadata.tokenUri === "string"
-        ? nftMetadata.tokenUri
-        : undefined,
+    tokenId: nftMetadata?.tokenId,
+    uri: nftMetadata?.tokenUri,
   };
   const { data: nft } = useNftMetadata(nftObject);
   const nftImage = useNftGlyphImage(nftObject);