+++ src/lib/components/decode-message/evm-message/decode-evm-message-erc20-approve.tsx
@@ -48,9 +48,9 @@ export const DecodeEvmMessageErc20ApproveHeader = ({
           <TokenImageWithAmount token={token} />
         </Flex>
         <Text color="text.dark">by</Text>
-        <ExplorerLink showCopyOnHover type="evm_tx_hash" value={from} />
+        <ExplorerLink showCopyOnHover type="user_address" value={from} />
         <Text color="text.dark">to</Text>
-        <ExplorerLink showCopyOnHover type="evm_tx_hash" value={spender} />
+        <ExplorerLink showCopyOnHover type="user_address" value={spender} />
       </DecodeMessageHeader>
     </Flex>
   );
@@ -68,7 +68,6 @@ export const DecodeEvmMessageErc20ApproveBody = ({
       isExpand
       log={undefined}
       sx={{
-        pb: 1,
         pl: 0,
       }}
     >
@@ -76,7 +75,7 @@ export const DecodeEvmMessageErc20ApproveBody = ({
         <ExplorerLink
           showCopyOnHover
           textFormat="normal"
-          type="evm_tx_hash"
+          type="user_address"
           value={from}
         />
       </DecodeMessageRow>
@@ -84,7 +83,7 @@ export const DecodeEvmMessageErc20ApproveBody = ({
         <ExplorerLink
           showCopyOnHover
           textFormat="normal"
-          type="evm_tx_hash"
+          type="user_address"
           value={spender}
         />
       </DecodeMessageRow>
@@ -93,7 +92,13 @@ export const DecodeEvmMessageErc20ApproveBody = ({
       </DecodeMessageRow>
       <DecodeMessageRow title="ERC-20 Contract">
         <ExplorerLink
-          leftIcon={<CustomIcon color="primary.main" name="contract-address" />}
+          leftIcon={
+            <CustomIcon
+              boxSize={4}
+              color="primary.main"
+              name="contract-address"
+            />
+          }
           showCopyOnHover
           textFormat="normal"
           type="evm_contract_address"