+++ src/lib/components/decode-message/evm-message/decode-evm-message-eth-transfer.tsx
@@ -47,9 +47,9 @@ export const DecodeEvmMessageEthTransferHeader = ({
           <TokenImageWithAmount token={token} />
         </Flex>
         <Text color="text.dark">from</Text>
-        <ExplorerLink showCopyOnHover type="evm_tx_hash" value={from} />
+        <ExplorerLink showCopyOnHover type="user_address" value={from} />
         <Text color="text.dark">to</Text>
-        <ExplorerLink showCopyOnHover type="evm_tx_hash" value={to} />
+        <ExplorerLink showCopyOnHover type="user_address" value={to} />
       </DecodeMessageHeader>
     </Flex>
   );
@@ -67,7 +67,6 @@ export const DecodeEvmMessageEthTransferBody = ({
       isExpand
       log={undefined}
       sx={{
-        pb: 1,
         pl: 0,
       }}
     >
@@ -75,7 +74,7 @@ export const DecodeEvmMessageEthTransferBody = ({
         <ExplorerLink
           showCopyOnHover
           textFormat="normal"
-          type="evm_tx_hash"
+          type="user_address"
           value={from}
         />
       </DecodeMessageRow>
@@ -83,7 +82,7 @@ export const DecodeEvmMessageEthTransferBody = ({
         <ExplorerLink
           showCopyOnHover
           textFormat="normal"
-          type="evm_tx_hash"
+          type="user_address"
           value={to}
         />
       </DecodeMessageRow>